+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.46.0] - 2026-09-03
+
 ### Added
 
 - **A syntax hint for JavaScript/TypeScript-style `for (const x of arr)` loops.**